Browserstack is fixed
The issue was related to executing an async function within the selenium context
