+++ package.json
@@ -1,15 +1,27 @@
 {
-  "description": "See the package.json at packages/upscaler for more information",
   "name": "@upscalerjs/root",
   "version": "1.0.0-beta.14",
   "private": true,
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/thekevinscott/UpscalerJS.git"
+  "babel": {
+    "env": {
+      "test": {
+        "plugins": [
+          "@babel/plugin-transform-modules-commonjs"
+        ]
+      }
+    },
+    "presets": [
+      [
+        "@babel/preset-env",
+        {
+          "targets": {
+            "node": "current"
+          }
+        }
+      ],
+      "@babel/preset-typescript"
+    ]
   },
-  "author": "Kevin Scott",
-  "license": "MIT",
-  "homepage": "https://github.com/thekevinscott/UpscalerJS",
   "scripts": {
     "build": "pnpm --filter=\"upscaler\" build",
     "dev:browser": "pnpm --filter=\"@upscalerjs/dev-browser\" run dev",
@@ -20,7 +32,6 @@
     "liteinstall": "pnpm --filter=!\"@upscalerjs/docs\" --filter=!\"@upscalerjs/examples.*\" install",
     "postinstall": "pnpm --filter upscaler run scaffold:dependencies 'browser' && pnpm --filter @upscalerjs/core --filter @upscalerjs/default-model build",
     "update:dependency": "pnpm --filter @upscalerjs/scripts update:dependency",
-    "update:npm:dependencies": "pnpm --filter @upscalerjs/scripts update:npm:dependencies",
     "update:tfjs": "pnpm --filter @upscalerjs/scripts update:tfjs",
     "update:version": "pnpm --filter @upscalerjs/scripts update:version && pnpm liteinstall",
     "commit-latest-version-changes": "git add packages/upscalerjs/src/constants.ts *examples/*/package.json* && git commit -m \"Updating CDN && example versions\" --no-verify",
@@ -48,112 +59,101 @@
     "test:unit:browser:jest": "pnpm --filter upscaler test:unit:browser:jest",
     "test:unit:node": "pnpm --filter upscaler test:unit:node"
   },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/thekevinscott/UpscalerJS.git"
+  },
+  "author": "Kevin Scott",
+  "license": "MIT",
+  "homepage": "https://github.com/thekevinscott/UpscalerJS",
   "peerDependencies": {
-    "@tensorflow/tfjs": "~4.8.0",
-    "@tensorflow/tfjs-node": "~4.8.0",
-    "@tensorflow/tfjs-node-gpu": "~4.8.0"
+    "@tensorflow/tfjs": "4.8.0",
+    "@tensorflow/tfjs-node": "4.8.0",
+    "@tensorflow/tfjs-node-gpu": "4.8.0"
   },
   "devDependencies": {
-    "@babel/cli": "^7.22.9",
-    "@babel/core": "^7.22.9",
-    "@babel/plugin-proposal-export-namespace-from": "^7.18.9",
-    "@babel/plugin-transform-modules-commonjs": "^7.22.5",
-    "@babel/preset-env": "^7.22.9",
-    "@babel/preset-typescript": "^7.22.5",
-    "@rollup/plugin-commonjs": "^25.0.3",
-    "@rollup/plugin-node-resolve": "^15.1.0",
-    "@schemastore/package": "^0.0.9",
-    "@tensorflow/tfjs": "~4.8.0",
-    "@tensorflow/tfjs-node": "~4.8.0",
-    "@tensorflow/tfjs-node-gpu": "~4.8.0",
-    "@types/babel__core": "^7.20.1",
-    "@types/chai": "^4.3.5",
-    "@types/chai-as-promised": "^7.1.5",
-    "@types/fs-extra": "^11.0.1",
-    "@types/inquirer": "^9.0.3",
-    "@types/jest": "^29.5.3",
-    "@types/jsdom": "^21.1.1",
-    "@types/mocha": "^10.0.1",
-    "@types/node": "^20.4.2",
-    "@types/pixelmatch": "^5.2.4",
-    "@types/pngjs": "^6.0.1",
-    "@types/selenium-webdriver": "^4.1.15",
-    "@types/serve-handler": "^6.1.1",
-    "@types/tar": "^6.1.5",
-    "@types/tiny-async-pool": "^2.0.0",
-    "@types/yargs": "^17.0.24",
-    "@typescript-eslint/eslint-plugin": "^6.1.0",
-    "@typescript-eslint/parser": "^6.1.0",
-    "babel-plugin-add-module-exports": "^1.0.4",
-    "browserstack-local": "^1.5.3",
-    "chai": "^4.3.7",
-    "chai-as-promised": "^7.1.1",
-    "chokidar": "^3.5.3",
-    "docsify-cli": "^4.4.4",
-    "dotenv": "^16.3.1",
-    "esbuild": "^0.18.15",
-    "eslint": "^8.45.0",
-    "eslint-config-prettier": "^8.8.0",
-    "eslint-plugin-jsdoc": "^46.4.4",
-    "eslint-plugin-prefer-arrow": "^1.2.3",
-    "fast-folder-size": "^2.1.0",
-    "fs-extra": "^11.1.1",
-    "heapdump": "^0.3.15",
-    "html-webpack-plugin": "^5.5.3",
-    "inquirer": "^9.2.8",
-    "jest": "^29.6.1",
-    "jest-canvas-mock": "^2.5.2",
-    "jsdom": "^22.1.0",
-    "mocha": "^10.2.0",
-    "pixelmatch": "^5.3.0",
-    "playwright-test": "^12.1.1",
-    "pngjs": "^7.0.0",
-    "prettier": "^3.0.0",
-    "puppeteer": "^19.0.0",
-    "rimraf": "^5.0.1",
-    "rollup": "^3.26.3",
+    "@babel/cli": "7.0.0-beta.44",
+    "@babel/core": "7.18.6",
+    "@babel/plugin-proposal-export-namespace-from": "7.18.9",
+    "@babel/plugin-transform-modules-commonjs": "7.18.6",
+    "@babel/preset-env": "7.0.0-beta.44",
+    "@babel/preset-typescript": "7.18.6",
+    "@rollup/plugin-commonjs": "22.0.1",
+    "@rollup/plugin-node-resolve": "13.3.0",
+    "@schemastore/package": "0.0.6",
+    "@tensorflow/tfjs": "4.8.0",
+    "@tensorflow/tfjs-node": "4.8.0",
+    "@tensorflow/tfjs-node-gpu": "4.8.0",
+    "@types/babel__core": "7.1.19",
+    "@types/chai": "4.3.1",
+    "@types/chai-as-promised": "7.1.5",
+    "@types/fs-extra": "9.0.13",
+    "@types/inquirer": "8.2.1",
+    "@types/jest": "28.1.4",
+    "@types/jsdom": "16.2.14",
+    "@types/mocha": "9.1.1",
+    "@types/node": "18.11.9",
+    "@types/pixelmatch": "5.2.4",
+    "@types/pngjs": "6.0.1",
+    "@types/puppeteer": "5.4.6",
+    "@types/rimraf": "3.0.2",
+    "@types/selenium-webdriver": "4.1.1",
+    "@types/serve-handler": "6.1.1",
+    "@types/tar": "6.1.1",
+    "@types/tiny-async-pool": "2.0.0",
+    "@types/yargs": "17.0.10",
+    "@typescript-eslint/eslint-plugin": "5.11.0",
+    "@typescript-eslint/parser": "5.42.1",
+    "babel-plugin-add-module-exports": "1.0.4",
+    "browserstack-local": "1.5.1",
+    "chai": "4.3.6",
+    "chai-as-promised": "7.1.1",
+    "chokidar": "3.5.3",
+    "docsify-cli": "4.4.4",
+    "dotenv": "16.0.1",
+    "esbuild": "0.11.0",
+    "eslint": "8.17.0",
+    "eslint-config-prettier": "8.3.0",
+    "eslint-plugin-jsdoc": "39.4.0",
+    "eslint-plugin-prefer-arrow": "1.2.3",
+    "fast-folder-size": "1.7.1",
+    "fs-extra": "10.1.0",
+    "heapdump": "0.3.15",
+    "inquirer": "8.2.4",
+    "jest": "29.0.0",
+    "jest-canvas-mock": "2.4.0",
+    "jsdom": "20.0.0",
+    "mocha": "10.0.0",
+    "pixelmatch": "5.3.0",
+    "playwright-test": "8.1.1",
+    "pngjs": "6.0.0",
+    "prettier": "2.7.1",
+    "puppeteer": "15.3.0",
+    "rimraf": "3.0.2",
+    "rollup": "2.75.7",
     "seedrandom": "3.0.5",
-    "selenium-webdriver": "^4.10.0",
-    "serve": "^14.2.0",
-    "serve-handler": "^6.1.5",
-    "tape": "^5.6.6",
-    "tar": "^6.1.15",
-    "tiny-async-pool": "^2.1.0",
-    "tree-kill": "^1.2.2",
-    "ts-jest": "^29.1.1",
-    "ts-node": "^10.9.1",
-    "typescript": "^5.1.6",
-    "uglify-js": "^3.17.4",
+    "selenium-webdriver": "4.3.0",
+    "serve": "13.0.2",
+    "serve-handler": "6.1.3",
+    "tape": "5.5.3",
+    "tar": "6.1.11",
+    "tiny-async-pool": "2.1.0",
+    "tree-kill": "1.2.2",
+    "ts-jest": "29.0.1",
+    "ts-node": "10.9.1",
+    "typescript": "4.8.3",
+    "uglify-js": "3.16.2",
     "upscaler": "workspace:*",
-    "vite": "^4.4.5",
-    "webpack": "^5.88.2",
-    "yargs": "^17.7.2"
-  },
-  "babel": {
-    "env": {
-      "test": {
-        "plugins": [
-          "@babel/plugin-transform-modules-commonjs"
-        ]
-      }
-    },
-    "presets": [
-      [
-        "@babel/preset-env",
-        {
-          "targets": {
-            "node": "current"
-          }
-        }
-      ],
-      "@babel/preset-typescript"
-    ]
+    "vite": "3.1.5",
+    "html-webpack-plugin": "5.5.0",
+    "webpack": "5.73.0",
+    "yargs": "17.6.2"
   },
   "pnpm": {
     "overrides": {
-      "@tensorflow/tfjs": "~4.8.0",
-      "@tensorflow/tfjs-node": "~4.8.0",
-      "@tensorflow/tfjs-node-gpu": "~4.8.0"
+      "@tensorflow/tfjs": "4.8.0",
+      "@tensorflow/tfjs-node": "4.8.0",
+      "@tensorflow/tfjs-node-gpu": "4.8.0"
     }
   }
 }
