+++ scripts/package-scripts/benchmark/performance/utils/utils.ts
@@ -88,58 +88,6 @@ export const runScript = async (cmd: string) => {
   return [stdout, stderr, err];
 };
 
-export async function executeAsyncScript<T>(driver: webdriver.WebDriver, fn: (args?: any) => Promise<T>, args?: any, {
-  pollTime = 100, timeout = 60 * 1000 * 5
-}: {
-  pollTime?: number;
-  timeout?: number;
-} = {}): Promise<T> {
-  const wait = (d: number) => new Promise(r => setTimeout(r, d));
-  const localKey = `___result_${Math.random()}___`;
-  const errorKey = `___result_${Math.random()}___`;
-  const mainFn = new Function(`
-    const main = ${fn.toString()}
-    main(...arguments).then((result) => {
-      window['${localKey}'] = result;
-    }).catch(err => {
-      window['${errorKey}'] = err.message;
-    });
-  `);
-  try {
-    driver.executeScript(mainFn, args);
-  } catch (err) {
-    if (err instanceof Error) {
-      throw new Error(`Error executing main script: ${err.message}`);
-    } else {
-      throw err;
-    }
-  }
-  let response: T | undefined;
-  let err: string | undefined;
-  const start = performance.now();
-  while (!response && !err) {
-    if (performance.now() - start > timeout) {
-      throw new Error(`Failed to execute script after ${timeout} ms`);
-    }
-    try {
-      response = await driver.executeScript<T | undefined>((localKey: string) => window[localKey], localKey);
-    } catch(err) {
-      console.error('Error executing script', err);
-    }
-    if (!response) {
-      err = await driver.executeScript<string | undefined>((errorKey: string) => window[errorKey], errorKey);
-      if (err) {
-        throw new Error(err);
-      }
-    }
-    await wait(pollTime);
-  }
-  if (!response) {
-    throw new Error('Bug with code');
-  }
-  return response;
-}
-
 declare global {
   interface Window {
     [index: string]: any;
